Mpesa/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==2.2
django-cors==0.1
django-cors-headers==3.2.1
django-mailgun==0.9.1
django-mysql==3.3.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.0.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.8
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
python-decouple==3.3
python-http-client==3.2.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.8
whitenoise==5.0.1
